+++ example/example_ogimet.py
@@ -5,7 +5,7 @@ Created on Fri Jan 24 22:34:11 2025
 @author: Kolja
 """
 from datetime import datetime
-from synop import OGIMET, SYNOP
+from synop import OGIMET, SYNOP_PARAM
         
 def main():
     
@@ -15,12 +15,12 @@ def main():
     stations.info()
     
     # test full year
-    s = SYNOP.block_full_year("10338", 2024)#.url()
+    s = SYNOP_PARAM.block_full_year("10338", 2024)#.url()
     table = o.request_synop(s, reload=False)
     table.info()
     
     # test full state    
-    s = SYNOP.state_full_day("Germany", datetime(2024,8,1))
+    s = SYNOP_PARAM.state_full_day("Germany", datetime(2024,8,1))
     table = o.request_synop(s, reload=False)
     table.info()